+++ Examples/AnalogMap/Windows/AnalogMap.c
@@ -52,28 +52,28 @@ static void TestFast(int32_t Value, int32_t InMin, int32_t InMax, int32_t OutMin
 
 extern int main(void)
 {
-  const long Min = 1;
-  const long Max = 1000;
+  const int32_t Min = 1;
+  const int32_t Max = 1000;
 
   //
   // Test accurate implementation...
   //
-  for (int I = Min; I <= Max; I++) {
+  for (unsigned I = Min; I <= Max; I++) {
     TestAccurate((double) I, (double) Min, (double) Max, 1.0, 25.0);
   }
 
-  for (int I = 1; I <= ITERATIONS; I++) {
+  for (unsigned I = 1; I <= ITERATIONS; I++) {
     printf("%d\n", AccurateDistribution[I]);
   }
 
   //
   // Test fast implementation...
   //
-  for (int I = Min; I <= Max; I++) {
+  for (int32_t I = Min; I <= Max; I++) {
     TestFast(I, Min, Max, 1, 25);
   }
 
-  for (int I = 1; I <= ITERATIONS; I++) {
+  for (unsigned I = 1; I <= ITERATIONS; I++) {
     printf("%d\n", FastDistribution[I]);
   }
